Tests/LibWeb/Text/expected/form-method-dialog.txt

  dialog.open before form submit: true
dialog.open after form submit: false
